requirements.txt
-e .
python-dotenv
pymupdf
pydantic
google-generativeai
google-genai
openai
pinecone
agno
duckduckgo-search
yfinance